app/be/auth/requirements/common.txt
connexion == 2.2.0
swagger-ui-bundle==0.0.3
Flask-SQLAlchemy==2.4.0
blinker==1.4.0
